requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django
psycopg2-binary==2.9.3
sqlparse==0.4.2
gunicorn==20.1.0
djangorestframework==3.13.1
bcrypt==4.0.0
django-cors-headers==3.13.0
PyJWT==2.4.0
requests==2.28.1
pysendpulse==0.1.4
python-decouple==3.6
python-jose-cryptodome
coverage
